+++ modular_skyrat/modules/admin_chat_icon/code/preference.dm
@@ -1,0 +1,5 @@
+/datum/preference/toggle/admin/ooc_admin_tag
+	category = PREFERENCE_CATEGORY_GAME_PREFERENCES
+	default_value = TRUE
+	savefile_key = "ooc_admin_tag"
+	savefile_identifier = PREFERENCE_PLAYER
