+++ tgstation.dme
@@ -4928,6 +4928,7 @@
 #include "modular_skyrat\modules\admin\code\manage_player_ranks.dm"
 #include "modular_skyrat\modules\admin\code\sooc.dm"
 #include "modular_skyrat\modules\admin\code\smites\pie.dm"
+#include "modular_skyrat\modules\admin_chat_icon\code\preference.dm"
 #include "modular_skyrat\modules\advanced_shuttles\code\areas.dm"
 #include "modular_skyrat\modules\advanced_shuttles\code\closet.dm"
 #include "modular_skyrat\modules\advanced_shuttles\code\computer.dm"
