--- tgui/packages/tgui/interfaces/PreferencesMenu/preferences/features/game_preferences/skyrat/auto_dementor.tsx
@@ -1,8 +1,0 @@
-import { FeatureToggle, CheckboxInput } from '../../base';
-
-export const auto_dementor_pref: FeatureToggle = {
-  name: 'Auto dementor',
-  category: 'ADMIN',
-  description: 'When enabled, you will automatically dementor.',
-  component: CheckboxInput,
-};
