--- tgui/packages/tgui/interfaces/PreferencesMenu/preferences/features/game_preferences/skyrat/delete_sparks.tsx
@@ -1,9 +1,0 @@
-import { CheckboxInput, FeatureToggle } from '../../base';
-
-export const delete_sparks_pref: FeatureToggle = {
-  name: 'Deletion sparks',
-  category: 'ADMIN',
-  description:
-    'Toggles if you want to play a sparking animation when deleting things as an admin.',
-  component: CheckboxInput,
-};
