--- tgui/packages/tgui/interfaces/PreferencesMenu/preferences/features/game_preferences/skyrat/looc.tsx
@@ -1,9 +1,0 @@
-import { CheckboxInput, FeatureToggle } from '../../base';
-
-export const looc_admin_pref: FeatureToggle = {
-  name: 'See admin LOOC',
-  category: 'ADMIN',
-  description:
-    'Toggles whether you want to see LOOC anywhere as an admin or not.',
-  component: CheckboxInput,
-};
