--- tgui/packages/tgui/interfaces/PreferencesMenu/preferences/features/game_preferences/skyrat/ticket_ping.tsx
@@ -1,9 +1,0 @@
-import { FeatureToggle, CheckboxInput } from '../../base';
-
-export const ticket_ping_pref: FeatureToggle = {
-  name: 'Ticket ping',
-  category: 'ADMIN',
-  description:
-    'When enabled, you will recieve regular pings from unhandled tickets.',
-  component: CheckboxInput,
-};
